chore: more test fixes
