+++ src/components/learner-credit-management/BudgetDetailCatalogTabContents.jsx
@@ -2,9 +2,9 @@ import React from 'react';
 import { Row, Col } from '@edx/paragon';
 
 const BudgetDetailCatalogTabContents = () => (
-  <Row>
+  <Row data-testid="budget-detail-catalog-tab-contents">
     <Col>
-      <h3>Catalog</h3>
+      <h3>Budget Name</h3>
       <p>TODO</p>
     </Col>
   </Row>
